requirements.txt
groq>=0.9.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-groq>=0.2.0
langgraph>=0.2.0
python-dotenv>=1.0.0
